Convert step wizard basics to TSX (WOR=1800).
